+++ packages/tide-predictor/src/constituents/3L2.ts
@@ -1,4 +1,4 @@
-import { defineCompoundConstituent } from "./index.js";
+import { defineCompoundConstituent } from "./definition.js";
 import L2 from "./L2.js";
 
 /**
