+++ packages/tide-predictor/src/constituents/MTM.ts
@@ -1,4 +1,4 @@
-import { defineCompoundConstituent } from "./index.js";
+import { defineCompoundConstituent } from "./definition.js";
 import M2 from "./M2.js";
 import T2 from "./T2.js";
