+++ packages/tide-predictor/src/constituents/MU2.ts
@@ -1,4 +1,4 @@
-import { defineCompoundConstituent } from "./index.js";
+import { defineCompoundConstituent } from "./definition.js";
 import M2 from "./M2.js";
 import S2 from "./S2.js";
