+++ packages/tide-predictor/src/constituents/RHO1.ts
@@ -1,4 +1,4 @@
-import { defineCompoundConstituent } from "./index.js";
+import { defineCompoundConstituent } from "./definition.js";
 import NU2 from "./NU2.js";
 import K1 from "./K1.js";
